[fix] 로그아웃시, user 정보가 없어지지 않는 버그 해결(로그아웃시 user 정보 null 추가) userStore + authStore 합침
