+++ src/stores/AuthStore.ts
@@ -1,11 +1,20 @@
-import firebase from 'firebase/compat/app';
 import { createContext } from 'react';
 import { makeObservable, observable } from 'mobx';
 import { getSignUp, getSignIn } from 'api/userAuth';
 
+import { onAuthStateChanged } from '@firebase/auth';
+import { authService } from 'service/firebase';
+
+interface IUser {
+  uid: string | null;
+  photoURL: string | null;
+  displayName: any;
+  email: string | null;
+  createdAt: any;
+}
+
 class AuthStore {
-  @observable user: firebase.User | null | firebase.auth.UserCredential | void =
-    null;
+  @observable user: IUser | null = null;
   @observable isLoggedIn: boolean = false;
 
   constructor() {
@@ -18,9 +27,8 @@ class AuthStore {
 
   signIn = async (email: string, password: string) => {
     const res = await getSignIn(email, password);
-    this.user = res;
 
-    if (this.user) {
+    if (res) {
       this.isLoggedIn = true;
     } else {
       this.isLoggedIn = false;
@@ -31,6 +39,20 @@ class AuthStore {
     this.user = null;
     this.isLoggedIn = false;
   };
+
+  unsubscribe = onAuthStateChanged(authService, user => {
+    if (user) {
+      const userObj = {
+        uid: user.uid,
+        photoURL: user.photoURL,
+        displayName:
+          user.displayName || authService.currentUser?.displayName || '카페인',
+        email: user.email,
+        createdAt: user.metadata.creationTime,
+      };
+      this.user = userObj;
+    }
+  });
 }
 
 export default createContext(new AuthStore());
